reserva 0.1

+++ apphotel/database/migrations/2024_03_02_010304_create_quartos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('numero');
             $table->string('tipo');
-            $table->decimal('valor', 8,2);
+            $table->double('valor', 8,2);
             $table->timestamps();
         });
     }
